chore(dashboard): move fetch method to utils
