+++ server/api/contributions-timeline.ts
@@ -1,6 +1,4 @@
-import { Query } from 'types/graphql-types'
-
-type User = Query['user']
+import { fetchContributionsCollectionAPI } from '@/server/utils'
 
 interface TimelineCommit {
   contributionCount: number
@@ -52,47 +50,6 @@ const query = `
     }
   }
 `
-
-/**
- * Fetch Github API and returns the user info
- * @param from from which date to fetch the data
- * @param to to the date to fetch the data
- * @returns the user object
- */
-async function fetchContributionsCollectionAPI(
-  from: Date,
-  to: Date
-): Promise<User> {
-  try {
-    const request = await fetch('https://api.github.com/graphql', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${process.env.NUXT_GITHUB_TOKEN}`,
-      },
-      body: JSON.stringify({
-        query,
-        variables: {
-          login: 'Draichi',
-          from,
-          to,
-        },
-      }),
-    })
-
-    const response = await request.json()
-
-    if (response.message) {
-      throw Error(response.message)
-    }
-
-    return response.data.user as User
-  } catch (error) {
-    console.error(error)
-
-    return {} as User
-  }
-}
 
 /**
  * Creates the base array with the month and the contributions
@@ -169,7 +126,7 @@ export default defineEventHandler(
 
         to.setFullYear(year)
 
-        const user = await fetchContributionsCollectionAPI(from, to)
+        const user = await fetchContributionsCollectionAPI(from, to, query)
 
         if (!user) {
           return {} as ContributionsCollection
